Make it work TODO : adjust UI and let `DrawableSaitenAdjustmentRuleset` use `NotePlayfield` only
